backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.9.2
PyYAML==6.0.2
jinja2==3.1.4
python-multipart==0.0.12
openapi-spec-validator==0.7.1
click==8.1.7
rich==13.7.0
requests>=2.31.0